[generator] Fixed all warnings (GCC 7.3.0)

+++ base/base_tests/stl_helpers_test.cpp
@@ -92,7 +92,7 @@ UNIT_TEST(LessBy)
     std::vector<Value> v = {{2, 2}, {0, 4}, {3, 1}, {4, 0}, {1, 3}};
     std::sort(v.begin(), v.end(), my::LessBy(&Value::first));
     for (size_t i = 0; i < v.size(); ++i)
-      TEST_EQUAL(i, v[i].first, ());
+      TEST_EQUAL(i, static_cast<size_t>(v[i].first), ());
 
     std::vector<Value const *> pv;
     for (auto const & p : v)
@@ -100,7 +100,7 @@ UNIT_TEST(LessBy)
 
     std::sort(pv.begin(), pv.end(), my::LessBy(&Value::second));
     for (size_t i = 0; i < pv.size(); ++i)
-      TEST_EQUAL(i, pv[i]->second, ());
+      TEST_EQUAL(i, static_cast<size_t>(pv[i]->second), ());
   }
 
   {
